feat: 파일 업로드 , 게시글 삭제 , api인스턴스 추가

+++ client/server/api/cms/board.delete.ts
@@ -13,11 +13,22 @@ export default defineEventHandler(async (event) => {
   }
 
   try {
-    const { error } = await client
+    // 첨부 파일 삭제
+    const { error: attachmentError } = await client
+      .from('TB_ATTACHMENT')
+      .update({ boardId: null })
+      .match({ boardId });
+
+    if (attachmentError) return { statusCode: 500, message: 'Internal Server Error' };
+
+    // 이제 게시글 삭제
+    const { data, error } = await client
       .from('TB_BOARD')
       .delete()
-      .in('boardId', [boardId]);
+      .eq('boardId', boardId)
+      .single();
 
+    console.log(data);
     if (error) throw error;
 
     return { statusCode: 200, message: '게시글을 삭제했습니다.' };
